Move DebugLevel enum to its own module
