+++ src/ferry_cli/helpers/supported_workflows/GetFilteredGroupInfo.py
@@ -2,10 +2,10 @@
 from typing import Any, Dict, List
 
 try:
-    from ferry_cli.helpers.auth import DebugLevel
+    from ferry_cli.helpers.debug_level import DebugLevel
     from ferry_cli.helpers.workflows import Workflow
 except ImportError:
-    from helpers.auth import DebugLevel  # type: ignore
+    from helpers.debug_level import DebugLevel  # type: ignore
     from helpers.workflows import Workflow  # type: ignore
